updated IAM examples
