+++ .doc_gen/metadata/iam_metadata.yaml
@@ -3355,7 +3355,7 @@ iam_Scenario_GroupBasics:
     iam: {AddUserToGroup, CreateGroup, CreateUser, CreateAccessKey, CreateRole, CreatePolicy, AttachRolePolicy, PutGroupPolicy,
           DeleteGroup, DeleteGroupPolicy, DeleteAccessKey, DeleteUser, RemoveUserFromGroup}
 iam_Scenario_CreateUserAssumeRole:
-  title: Learn core operations for &IAM-user; using an &AWS;. 
+  title: Learn core operations for &IAM; using an &AWS; SDK. 
   title_abbrev: Create a user and assume a role
   synopsis: create a user and assume a role. &aws_sec_sdk_use-federation-warning;
   synopsis_list:
